move MapperLimits into colors.py
